feat: Add deployment instructions for Heroku (#16)
The code changes include a new section in the documentation that provides step-by-step instructions for deploying the project to Heroku. The instructions cover creating a Heroku app, enabling container registry, setting up PostgreSQL and Redis addons, configuring environment variables, pushing the image to the registry, releasing the image to the app, checking the logs, and opening the app in the browser.

Recent user commits:
- feat: geodjango deps and setup with postgis
- chore: update TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN, and TWILIO_SERVICE_SID environment variables
- chore: update TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN, and TWILIO_SERVICE_SID environment variables
- test_user_can_log_in: mock twilio_client.Client in AuthenticationTest
- chore: fix tests to path login check code methods
- fix: test to use relative imports
- feat: update CircleCI configuration to use the latest Ubuntu image
- lint
- lint
- feat: celery redis and twilio